# services/api-service/internal/swaggerui/openapi.yaml
openapi: 3.0.3
info:
  title: ASOC Public API (api-service)
  description: |
    **api-service.** `POST /api/v1/scans` — единый вход: в теле задаётся **`scanner_id`** (`semgrep`, `gitleaks`, …), далее тот же сценарий (сканирование, ingest находок, группы, тикеты).
    `POST /api/v1/scans/semgrep`, `…/gitleaks`, `…/sca` (Trivy) и `…/dast` (OWASP ZAP) оставлены для совместимости (тело как при соответствующем сканере — без поля `scanner_id`).
    **`POST /api/v1/findings/ingest`** — приём **уже нормализованных** находок в формате ниже (CI, сторонний сканер); поле `owner_user_id` в теле **не используется**, владелец берётся только из JWT пользователя консоли при наличии. Поле **`console_product_id`** (опционально) привязывает прогон к записи **`GET /api/v1/console/products`** того же пользователя; без JWT отбрасывается. Отчёт и группы по одному проекту — **`GET /api/v1/report/vulnerabilities?console_product_id=…`** и **`GET /api/v1/groups?console_product_id=…`** (нужен JWT пользователя, не только API-ключ). С внешнего контура авторизуйте запрос тем же **`X-API-Key` / Bearer API-ключа** или JWT (см. `APP_AUTH_API_KEY`).
    `GET /api/v1/integrations` — публичный каталог сканеров (без секретов): метаданные, флаг доступности для консоли.
    Способ передачи находок и настройки среды — в `docs/ENVIRONMENT.md` и `docs/ARCHITECTURE.md`.
  version: "1.0"
servers:
  - url: https://atomic-asoc.ru
    description: Стенд (через публичный вход / asoc-web)
  - url: http://localhost:8080
...
paths:
...
  /api/v1/findings/ingest:
    post:
      summary: Ingest находок произвольного сканера (нормализованный JSON)
      description: >
        Отправка пакета в тот же контракт, что internal `processing-service`; маршрутизация через api-service (Kafka или HTTP по ENV).
        Пустой массив `findings` допустим — создаётся run без строк.
        Поле `channel` (manual/ci): если не указано — подставляется **ci** (ожидаемый сценарий из CI/CD).
      operationId: ingestFindingsPublic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessingIngestRequestPublic"
      responses:
        "202":
          description: Пакет принят (Kafka); при брокере — status/correlation_id, без run_id
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/IngestAccepted"
                  - $ref: "#/components/schemas/ProcessingSummary"
        "400":
          description: Неверное тело
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Нет API-ключа / JWT для /api/*
        "502":
          description: Ошибка Kafka / processing-service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
...
components:
  schemas:
...
    ErrorBody:
      type: object
      properties:
        error:
          type: string
    ProcessingFindingItem:
      type: object
      required:
        - asset_id
        - identifier
        - severity
        - component
        - version
        - cve
        - cwe
      properties:
        asset_id:
          type: string
        identifier:
          type: string
        severity:
          type: string
        component:
          type: string
        version:
          type: string
        cve:
          type: string
        cwe:
          type: string
        metadata:
          type: object
          additionalProperties: true
        raw_payload:
          type: object
          additionalProperties: true
    ProcessingIngestRequestPublic:
      type: object
      required:
        - scanner_name
        - findings
      properties:
        scanner_name:
          type: string
          description: Произвольное имя источника (отображается в processing run).
          example: my-custom-sast
        findings:
          type: array
          items:
            $ref: "#/components/schemas/ProcessingFindingItem"
        owner_user_id:
          type: integer
          format: int64
          deprecated: true
          description: Игнорируется; владелец задаётся только JWT консоли на api-service.
        channel:
          type: string
          enum: [manual, ci]
          description: >
            Канал прогона. Если не указано, для POST /api/v1/findings/ingest подставляется «ci» (типичный сценарий пайплайна).
            Ручные сканы через POST /api/v1/scans всегда фиксируются как «manual» на стороне сервера.
          example: ci
        console_product_id:
          type: integer
          format: int64
          description: >
            Опционально: id строки из GET /api/v1/console/products (продукт того же пользователя JWT).
            Прогон сохраняется с привязкой к проекту; GET /api/v1/groups и GET /api/v1/report/vulnerabilities с query-параметром
            console_product_id возвращают только данные прогонов этого продукта. С одним API-ключом без JWT поле отбрасывается.
...
    IngestAccepted:
      type: object
      properties:
        status:
          type: string
          example: accepted
        correlation_id:
          type: string
          format: uuid
        findings_queued:
          type: integer
    ProcessingSummary:
      type: object
      properties:
        run_id:
          type: integer
          format: int64
        findings_received:
          type: integer
        findings_processed:
          type: integer
        vulnerabilities_created:
          type: integer
        groups_updated:
          type: integer
...